audio_safety_detection/requirements.txt
torch>=2.1.0
torchaudio>=2.1.0
transformers>=4.35.0
faster-whisper>=0.10.0
openai-whisper>=20231117
detoxify>=0.5.2
librosa>=0.10.0
soundfile>=0.12.1
ffmpeg-python>=0.2.0
accelerate>=0.24.0
optimum>=1.15.0
numpy>=1.24.0
scipy>=1.11.0
pydantic>=2.5.0
loguru>=0.7.2
pyyaml>=6.0.1
tqdm>=4.66.0
psutil>=5.9.0
moviepy>=1.0.3